rename: SearchBar 폴더 생성으로 인한 경로 수정
